tests: remove duplicated code for plugin commands
