+++ tests/console/commands/plugin/test_remove.py
@@ -3,43 +3,12 @@ import tomlkit
 
 from poetry.__version__ import __version__
 from poetry.core.packages.package import Package
-from poetry.factory import Factory
 from poetry.layouts.layout import POETRY_DEFAULT
-from poetry.repositories.installed_repository import InstalledRepository
-from poetry.repositories.pool import Pool
-from poetry.utils.env import EnvManager
 
 
 @pytest.fixture()
 def tester(command_tester_factory):
     return command_tester_factory("plugin remove")
-
-
-@pytest.fixture()
-def installed():
-    repository = InstalledRepository()
-
-    repository.add_package(Package("poetry", __version__))
-
-    return repository
-
-
-def configure_sources_factory(repo):
-    def _configure_sources(poetry, sources, config, io):
-        pool = Pool()
-        pool.add_repository(repo)
-        poetry.set_pool(pool)
-
-    return _configure_sources
-
-
-@pytest.fixture(autouse=True)
-def setup_mocks(mocker, env, repo, installed):
-    mocker.patch.object(EnvManager, "get_system_env", return_value=env)
-    mocker.patch.object(InstalledRepository, "load", return_value=installed)
-    mocker.patch.object(
-        Factory, "configure_sources", side_effect=configure_sources_factory(repo)
-    )
 
 
 @pytest.fixture()
@@ -60,7 +29,8 @@ def pyproject(env):
     )
 
 
-def test_remove_installed_package(app, repo, tester, env, installed, pyproject):
+@pytest.fixture(autouse=True)
+def install_plugin(env, installed, pyproject):
     lock_content = {
         "package": [
             {
@@ -99,6 +69,8 @@ def test_remove_installed_package(app, repo, tester, env, installed, pyproject):
 
     installed.add_package(Package("poetry-plugin", "1.2.3"))
 
+
+def test_remove_installed_package(app, tester, env):
     tester.execute("poetry-plugin")
 
     expected = """\
@@ -124,45 +96,7 @@ Package operations: 0 installs, 0 updates, 1 removal
     assert "poetry-plugin" not in content["dependencies"]
 
 
-def test_remove_installed_package_dry_run(app, repo, tester, env, installed, pyproject):
-    lock_content = {
-        "package": [
-            {
-                "name": "poetry-plugin",
-                "version": "1.2.3",
-                "category": "main",
-                "optional": False,
-                "platform": "*",
-                "python-versions": "*",
-                "checksum": [],
-            },
-        ],
-        "metadata": {
-            "python-versions": "^3.6",
-            "platform": "*",
-            "content-hash": "123456789",
-            "hashes": {"poetry-plugin": []},
-        },
-    }
-
-    env.path.joinpath("poetry.lock").write_text(
-        tomlkit.dumps(lock_content), encoding="utf-8"
-    )
-
-    pyproject = tomlkit.loads(
-        env.path.joinpath("pyproject.toml").read_text(encoding="utf-8")
-    )
-    content = pyproject["tool"]["poetry"]
-
-    dependency_section = content["dependencies"]
-    dependency_section["poetry-plugin"] = "^1.2.3"
-
-    env.path.joinpath("pyproject.toml").write_text(
-        tomlkit.dumps(pyproject), encoding="utf-8"
-    )
-
-    installed.add_package(Package("poetry-plugin", "1.2.3"))
-
+def test_remove_installed_package_dry_run(app, tester, env):
     tester.execute("poetry-plugin --dry-run")
 
     expected = """\
